+++ src/scrape/scrape_linkedin.py
@@ -102,9 +102,3 @@ class LinkedinSpider(scrapy.Spider):
             writer = csv.writer(f)
             writer.writerow(list(result.values()))
         # yield result["url"]
-
-
-def scrape():
-    process = CrawlerProcess()
-    process.crawl(LinkedinSpider)
-    process.start()
