+++ src/sweify/__main__.py
@@ -2,15 +2,21 @@ import sys
 import time
 
 import schedule
-from scrape import scrape_linkedin
-from scrape.scrape_linkedin import scrape
+from scrape.scrape_computerjobs import ComputerjobsSpider
+from scrape.scrape_linkedin import LinkedinSpider
+from scrapy.crawler import CrawlerProcess
 from transform import parse
 
 
-def etl():
-    scrape_linkedin.scrape()
-    parse(upload=True)
+def scrape():
+    spider = CrawlerProcess()
+    spider.crawl(LinkedinSpider)
+    spider.crawl(ComputerjobsSpider)
+    spider.start()
 
+def etl():
+    scrape()
+    parse(upload=True)
 
 if __name__ == "__main__":
     if len(sys.argv) == 1:
@@ -20,9 +26,9 @@ if __name__ == "__main__":
             schedule.run_pending()
             time.sleep(1)
 
-    if sys.argv[1] == "--scrape-only":
+    if "--scrape-only" in sys.argv:
         scrape()
-    elif sys.argv[1] == "--parse-only":
+    elif "--parse-only" in sys.argv:
         if "--upload" in sys.argv:
             parse(upload=True)
         else:
